Added use of RoomManager in WebSocketController.

+++ backend/Battle/IRoomManager.cs
@@ -1,6 +1,6 @@
 ﻿namespace backend.Battle;
 public interface IRoomManager {
-    bool Push(int newScore, Room room);
+    bool Push(float newScore, Room room);
     
     Room? Pop();
     Room? GetRoom(int roomId);
